Refactored POFieldStyle
